Use cuDNN routine FindEx to find best algorithm.
FindEx is more stable than Get (heuristic-based) because it runs all the available algorithms and sorts them according to their speed.
Both Get and FindEx are supported now and can be specified through the definition of each layer in prototxt.
In Reshape, check whether shape of bottom and convolution descriptors have changed.
In caffe time, do multiple (instead of one) fwd/bwd pass in the initilization phase.
This is crucial because FindEx is executed in the first iterations and it takes quite a long time.